サンプルにecce実行できるものにしてをbot.jsに移行
