+++ builder/bot.js
@@ -1,53 +1,111 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-const { ActivityHandler, MessageFactory } = require('botbuilder');
+const { ActivityHandler } = require('botbuilder');
 const fetch = require('node-fetch');
 
-class EchoBot extends ActivityHandler {
-    constructor() {
+// The accessor names for the conversation data and user profile state property accessors.
+const CONVERSATION_DATA_PROPERTY = 'conversationData';
+const USER_PROFILE_PROPERTY = 'userProfile';
+
+class StateManagementBot extends ActivityHandler {
+    constructor(conversationState, userState) {
         super();
-        // See https://aka.ms/about-bot-activity-message to learn more about the message and other activity types.
-        this.onMessage(async (context, next) => {
-            await fetch('https://api.rinna.co.jp/models/ecce', {
-                method: 'POST',
-                headers: {
-                    'Content-type': 'application/json',
-                    'Cache-Control': 'no-cache',
-                    'Ocp-Apim-Subscription-Key': process.env.RINNA_SUBSCRIPTIONKEY_ECCE
-                },
-                body: JSON.stringify({
-                    'knowledgePath': 'ECCE_Sample.txt',
-                    'query': context.activity.text,
-                    'dialogHistory': [],
-                    'l2ReturnNum': 3,
-                    'l3ReturnNum': 1
+        // Create the state property accessors for the conversation data and user profile.
+        this.conversationDataAccessor = conversationState.createProperty(CONVERSATION_DATA_PROPERTY);
+        this.userProfileAccessor = userState.createProperty(USER_PROFILE_PROPERTY);
+
+        // The state management objects for the conversation and user state.
+        this.conversationState = conversationState;
+        this.userState = userState;
+
+        this.onMessage(async (turnContext, next) => {
+            // Get the state properties from the turn context.
+            const userProfile = await this.userProfileAccessor.get(turnContext, {});
+            const conversationData = await this.conversationDataAccessor.get(
+                turnContext, { promptedForUserName: false });
+
+            if (!userProfile.name) {
+                // First time around this is undefined, so we will prompt user for name.
+                if (conversationData.promptedForUserName) {
+                    // Set the name to what the user provided.
+                    userProfile.name = turnContext.activity.text;
+
+                    // Acknowledge that we got their name.
+                    await turnContext.sendActivity(`Thanks ${ userProfile.name }. To see conversation data, type anything.`);
+
+                    // Reset the flag to allow the bot to go though the cycle again.
+                    conversationData.promptedForUserName = false;
+                } else {
+                    // Prompt the user for their name.
+                    await turnContext.sendActivity('What is your name?');
+
+                    // Set the flag to true, so we don't prompt in the next turn.
+                    conversationData.promptedForUserName = true;
+                }
+            } else {
+                // Add message details to the conversation data.
+                conversationData.timestamp = turnContext.activity.timestamp.toLocaleString();
+                conversationData.channelId = turnContext.activity.channelId;
+
+                // Display state data.
+                await turnContext.sendActivity(`${ userProfile.name } sent: ${ turnContext.activity.text }`);
+                await turnContext.sendActivity(`Message received at: ${ conversationData.timestamp }`);
+                await turnContext.sendActivity(`Message received from: ${ conversationData.channelId }`);
+
+                await fetch('https://api.rinna.co.jp/models/ecce', {
+                    method: 'POST',
+                    headers: {
+                        'Content-type': 'application/json',
+                        'Cache-Control': 'no-cache',
+                        'Ocp-Apim-Subscription-Key': process.env.RINNA_SUBSCRIPTIONKEY_ECCE
+                    },
+                    body: JSON.stringify({
+                        'knowledgePath': 'ECCE_Sample.txt',
+                        'query': turnContext.activity.text,
+                        'dialogHistory': [],
+                        'l2ReturnNum': 3,
+                        'l3ReturnNum': 1
+                    })
                 })
-            })
-                .then(response => response.text())
-                .then(async (result) => {
-                    console.log(result);
-                    const data = JSON.parse(result);
-                    const replyText = data.resultResponseText;
-                    await context.sendActivity(MessageFactory.text(replyText, replyText));
-                    // By calling next() you ensure that the next BotHandler is run.
-                    await next();
-                })
-                .catch(error => console.log('error', error));
+                    .then(response => response.text())
+                    .then(async (result) => {
+                        console.log(result);
+                        const data = JSON.parse(result);
+                        const replyText = data.resultResponseText;
+                        await turnContext.sendActivity(replyText);
+                        // By calling next() you ensure that the next BotHandler is run.
+                        await next();
+                    })
+                    .catch(error => console.log('error', error));
+            }
+
+            // By calling next() you ensure that the next BotHandler is run.
+            await next();
         });
 
         this.onMembersAdded(async (context, next) => {
             const membersAdded = context.activity.membersAdded;
-            const welcomeText = 'Hello and welcome!';
             for (let cnt = 0; cnt < membersAdded.length; ++cnt) {
                 if (membersAdded[cnt].id !== context.activity.recipient.id) {
-                    await context.sendActivity(MessageFactory.text(welcomeText, welcomeText));
+                    await context.sendActivity('Welcome to State Bot Sample. Type anything to get started.');
                 }
             }
             // By calling next() you ensure that the next BotHandler is run.
             await next();
         });
     }
+
+    /**
+     * Override the ActivityHandler.run() method to save state changes after the bot logic completes.
+     */
+    async run(context) {
+        await super.run(context);
+
+        // Save any state changes. The load happened during the execution of the Dialog.
+        await this.conversationState.saveChanges(context, false);
+        await this.userState.saveChanges(context, false);
+    }
 }
 
-module.exports.EchoBot = EchoBot;
+module.exports.StateManagementBot = StateManagementBot;
