--- builder/bots/stateManagementBot.js
@@ -1,84 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-const { ActivityHandler } = require('botbuilder');
-
-// The accessor names for the conversation data and user profile state property accessors.
-const CONVERSATION_DATA_PROPERTY = 'conversationData';
-const USER_PROFILE_PROPERTY = 'userProfile';
-
-class StateManagementBot extends ActivityHandler {
-    constructor(conversationState, userState) {
-        super();
-        // Create the state property accessors for the conversation data and user profile.
-        this.conversationDataAccessor = conversationState.createProperty(CONVERSATION_DATA_PROPERTY);
-        this.userProfileAccessor = userState.createProperty(USER_PROFILE_PROPERTY);
-
-        // The state management objects for the conversation and user state.
-        this.conversationState = conversationState;
-        this.userState = userState;
-
-        this.onMessage(async (turnContext, next) => {
-            // Get the state properties from the turn context.
-            const userProfile = await this.userProfileAccessor.get(turnContext, {});
-            const conversationData = await this.conversationDataAccessor.get(
-                turnContext, { promptedForUserName: false });
-
-            if (!userProfile.name) {
-                // First time around this is undefined, so we will prompt user for name.
-                if (conversationData.promptedForUserName) {
-                    // Set the name to what the user provided.
-                    userProfile.name = turnContext.activity.text;
-
-                    // Acknowledge that we got their name.
-                    await turnContext.sendActivity(`Thanks ${ userProfile.name }. To see conversation data, type anything.`);
-
-                    // Reset the flag to allow the bot to go though the cycle again.
-                    conversationData.promptedForUserName = false;
-                } else {
-                    // Prompt the user for their name.
-                    await turnContext.sendActivity('What is your name?');
-
-                    // Set the flag to true, so we don't prompt in the next turn.
-                    conversationData.promptedForUserName = true;
-                }
-            } else {
-                // Add message details to the conversation data.
-                conversationData.timestamp = turnContext.activity.timestamp.toLocaleString();
-                conversationData.channelId = turnContext.activity.channelId;
-
-                // Display state data.
-                await turnContext.sendActivity(`${ userProfile.name } sent: ${ turnContext.activity.text }`);
-                await turnContext.sendActivity(`Message received at: ${ conversationData.timestamp }`);
-                await turnContext.sendActivity(`Message received from: ${ conversationData.channelId }`);
-            }
-
-            // By calling next() you ensure that the next BotHandler is run.
-            await next();
-        });
-
-        this.onMembersAdded(async (context, next) => {
-            const membersAdded = context.activity.membersAdded;
-            for (let cnt = 0; cnt < membersAdded.length; ++cnt) {
-                if (membersAdded[cnt].id !== context.activity.recipient.id) {
-                    await context.sendActivity('Welcome to State Bot Sample. Type anything to get started.');
-                }
-            }
-            // By calling next() you ensure that the next BotHandler is run.
-            await next();
-        });
-    }
-
-    /**
-     * Override the ActivityHandler.run() method to save state changes after the bot logic completes.
-     */
-    async run(context) {
-        await super.run(context);
-
-        // Save any state changes. The load happened during the execution of the Dialog.
-        await this.conversationState.saveChanges(context, false);
-        await this.userState.saveChanges(context, false);
-    }
-}
-
-module.exports.StateManagementBot = StateManagementBot;
